+++ pytdx/hq.py
@@ -11,7 +11,8 @@ import sys
 import pandas as pd
 
 if __name__ == '__main__':
-    sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+    sys.path.append(os.path.dirname(
+        os.path.dirname(os.path.realpath(__file__))))
 
 from pytdx.log import DEBUG, log
 from pytdx.parser.get_security_bars import GetSecurityBarsCmd
@@ -27,12 +28,14 @@ from pytdx.parser.get_company_info_category import GetCompanyInfoCategory
 from pytdx.parser.get_company_info_content import GetCompanyInfoContent
 from pytdx.parser.get_xdxr_info import GetXdXrInfo
 from pytdx.parser.get_finance_info import GetFinanceInfo
-from pytdx.util import get_real_trade_date,trade_date_sse
+from pytdx.parser.get_block_info import GetBlockInfo, GetBlockInfoMeta, get_and_parse_block_info
+from pytdx.util import get_real_trade_date, trade_date_sse
 from pytdx.params import TDXParams
 from pytdx.heartbeat import HqHeartBeatThread
 
 from pytdx.parser.setup_commands import SetupCmd1, SetupCmd2, SetupCmd3
-import threading,datetime
+import threading
+import datetime
 import random
 
 from pytdx.base_socket_client import BaseSocketClient, update_last_ack_time
@@ -45,8 +48,9 @@ class TdxHq_API(BaseSocketClient):
         SetupCmd2(self.client).call_api()
         SetupCmd3(self.client).call_api()
 
-    #### API List
+    # API List
 
+    # Notice：，如果一个股票当天停牌，那天的K线还是能取到，成交量为0
     @update_last_ack_time
     def get_security_bars(self, category, market, code, start, count):
         cmd = GetSecurityBarsCmd(self.client, lock=self.lock)
@@ -125,28 +129,48 @@ class TdxHq_API(BaseSocketClient):
         cmd.setParams(market, code)
         return cmd.call_api()
 
+
+    @update_last_ack_time
+    def get_block_info_meta(self, blockfile):
+        cmd = GetBlockInfoMeta(self.client, lock=self.lock)
+        cmd.setParams(blockfile)
+        return cmd.call_api()
+
+    @update_last_ack_time
+    def get_block_info(self, blockfile, start, size):
+        cmd = GetBlockInfo(self.client, lock=self.lock)
+        cmd.setParams(blockfile, start, size)
+        return cmd.call_api()
+
+    def get_and_parse_block_info(self, blockfile):
+        return get_and_parse_block_info(self, blockfile)
+
+
     def do_heartbeat(self):
         self.get_security_count(random.randint(0, 1))
 
-
-    def get_k_data(self, code, start,end):
+    def get_k_data(self, code, start_date, end_date):
         # 具体详情参见 https://github.com/rainx/pytdx/issues/5
-        if str(code)[0]=='6':
-            #0 - 深圳， 1 - 上海
-            market_code=1
-        else:
-            market_code=0
-        start_date=get_real_trade_date(start,1)
-        end_date=get_real_trade_date(end,-1)
-        index_0=str(datetime.date.today())
-        index_of_index_0=trade_date_sse.index(index_0)
-        index_of_index_end=trade_date_sse.index(end_date)
-        index_of_index_start=trade_date_sse.index(start_date)
-        
-        index_of_end=index_of_index_0-index_of_index_end
-        index_length=index_of_index_end+1-index_of_index_start
-        return self.get_security_bars(9, market_code, code,index_of_end, index_length)  # 返回普通list
-        
+        # 具体详情参见 https://github.com/rainx/pytdx/issues/21
+
+        # 新版一劳永逸偷懒写法zzz
+        market_code = 1 if str(code)[0] == '6' else 0
+        #https://github.com/rainx/pytdx/issues/33
+        # 0 - 深圳， 1 - 上海
+        start_date = get_real_trade_date(start_date, 1)
+        end_date = get_real_trade_date(end_date, -1)
+        data = []
+        for i in range(10):
+            data += self.get_security_bars(9, 0, code, (9 - i) * 800, 800)
+
+        data = self.to_df(data)
+        data['date'] = data['datetime'].apply(lambda x: x[0:10])
+        data['date'] = pd.to_datetime(data['date'])
+        data = data.set_index('date')
+        data = data.drop(['year', 'month', 'day', 'hour',
+                          'minute', 'datetime'], axis=1)
+        return data[start_date:end_date]
+
 
 if __name__ == '__main__':
     import pprint
@@ -157,7 +181,7 @@ if __name__ == '__main__':
         stocks = api.get_security_quotes([(0, "000001"), (1, "600300")])
         pprint.pprint(stocks)
         log.info("获取k线")
-        data = api.get_security_bars(9,0, '000001', 4, 3)
+        data = api.get_security_bars(9, 0, '000001', 4, 3)
         pprint.pprint(data)
         log.info("获取 深市 股票数量")
         pprint.pprint(api.get_security_count(0))
@@ -165,19 +189,21 @@ if __name__ == '__main__':
         stocks = api.get_security_list(1, 255)
         pprint.pprint(stocks)
         log.info("获取指数k线")
-        data = api.get_index_bars(9,1, '000001', 1, 2)
+        data = api.get_index_bars(9, 1, '000001', 1, 2)
         pprint.pprint(data)
         log.info("查询分时行情")
         data = api.get_minute_time_data(TDXParams.MARKET_SH, '600300')
         pprint.pprint(data)
         log.info("查询历史分时行情")
-        data = api.get_history_minute_time_data(TDXParams.MARKET_SH, '600300', 20161209)
+        data = api.get_history_minute_time_data(
+            TDXParams.MARKET_SH, '600300', 20161209)
         pprint.pprint(data)
         log.info("查询分时成交")
         data = api.get_transaction_data(TDXParams.MARKET_SZ, '000001', 0, 30)
         pprint.pprint(data)
         log.info("查询历史分时成交")
-        data = api.get_history_transaction_data(TDXParams.MARKET_SZ, '000001', 0, 10, 20170209)
+        data = api.get_history_transaction_data(
+            TDXParams.MARKET_SZ, '000001', 0, 10, 20170209)
         pprint.pprint(data)
         log.info("查询公司信息目录")
         data = api.get_company_info_category(TDXParams.MARKET_SZ, '000001')
@@ -192,10 +218,7 @@ if __name__ == '__main__':
         data = api.get_finance_info(0, '000001')
         pprint.pprint(data)
         log.info("日线级别k线获取函数")
-        data =api.get_k_data('000001','2017-07-01','2017-07-10')
+        data = api.get_k_data('000001', '2017-07-01', '2017-07-10')
         pprint.pprint(data)
 
         api.disconnect()
-
-
-