+++ pytdx/params.py
@@ -30,6 +30,7 @@ class TDXParams:
     KLINE_TYPE_DAILY = 4
     KLINE_TYPE_WEEKLY = 5
     KLINE_TYPE_MONTHLY = 6
+    KLINE_TYPE_EXHQ_1MIN = 7
     KLINE_TYPE_1MIN = 8
     KLINE_TYPE_RI_K = 9
     KLINE_TYPE_3MONTH = 10
@@ -41,3 +42,10 @@ class TDXParams:
     MAX_TRANSACTION_COUNT = 2000
     # k先数据最多800条
     MAX_KLINE_COUNT = 800
+
+
+    # 板块相关参数
+    BLOCK_SZ = "block_zs.dat"
+    BLOCK_FG = "block_fg.dat"
+    BLOCK_GN = "block_gn.dat"
+    BLOCK_DEFAULT = "block.dat"