+++ pytdx/parser/ex_get_instrument_quote.py
@@ -134,6 +134,12 @@ class GetInstrumentQuote(BaseParser):
         ]
 
 
+if __name__ == '__main__':
+    from pytdx.exhq import TdxExHq_API
+
+    api = TdxExHq_API()
+    with api.connect('61.152.107.141', 7727):
+        print(api.to_df(api.get_instrument_quote(47, "IF1709")))