+++ pytdx/parser/get_xdxr_info.py
@@ -17,6 +17,43 @@ get_volume ?
 1218.0031249523163 ---> 2.3
 
 """
+"""
+
+        1 除权除息 002175 2008-05-29
+        2 送配股上市  000656 2015-04-29
+        3 非流通股上市 000656 2010-02-10
+        4 未知股本变动 600642 1993-07-19
+        5 股本变化 000656 2017-06-30
+        6 增发新股 600887 2002-08-20
+        7 股份回购  600619 2000-09-08
+        8 增发新股上市 600186 2001-02-14
+        9 转配股上市 600811 2017-07-25
+        10 可转债上市 600418 2006-07-07
+        11 扩缩股  600381 2014-06-27
+        12 非流通股缩股 600339 2006-04-10
+        13 送认购权证 600008 2006-04-19
+        14 送认沽权证 000932 2006-03-01
+
+"""
+
+
+XDXR_CATEGORY_MAPPING = {
+    1 : "除权除息",
+    2 : "送配股上市",
+    3 : "非流通股上市",
+    4 : "未知股本变动",
+    5 : "股本变化",
+    6 : "增发新股",
+    7 : "股份回购",
+    8 : "增发新股上市",
+    9 : "转配股上市",
+    10 : "可转债上市",
+    11 : "扩缩股",
+    12 : "非流通股缩股",
+    13 : "送认购权证",
+    14 : "送认沽权证"
+}
+
 
 class GetXdXrInfo(BaseParser):
 
@@ -38,6 +75,13 @@ class GetXdXrInfo(BaseParser):
         pos += 2
 
         rows = []
+
+        def _get_v(v):
+            if v == 0:
+                return 0
+            else:
+                return get_volume(v)
+
         for i in range(num):
             market, code = struct.unpack(u"<B6s", body_buf[:7])
             pos += 7
@@ -53,15 +97,26 @@ class GetXdXrInfo(BaseParser):
             # b'\x00\xc0\x0fF' => 9200.00000
             # b'\x00@\x83E' => 4200.0000
 
-            panqianliutong_raw, qianzongguben_raw, panhouliutong_raw, houzongguben_raw = struct.unpack("<IIII", body_buf[pos: pos + 16])
+            suogu = None
+            panqianliutong, panhouliutong, qianzongguben, houzongguben = None, None, None, None
+            songzhuangu, fenhong, peigu, peigujia = None, None, None, None
+            fenshu, xingquanjia = None, None
+            if category == 1:
+                fenhong, peigujia, songzhuangu, peigu  = struct.unpack("<ffff", body_buf[pos: pos + 16])
+            elif category in [11, 12]:
+                (_, _, suogu, _) = struct.unpack("<IIfI", body_buf[pos: pos + 16])
+            elif category in [13, 14]:
+                xingquanjia, _, fenshu, _ = struct.unpack("<fIfI", body_buf[pos: pos + 16])
+            else:
+                panqianliutong_raw, qianzongguben_raw, panhouliutong_raw, houzongguben_raw = struct.unpack("<IIII", body_buf[pos: pos + 16])
+                panqianliutong = _get_v(panqianliutong_raw)
+                panhouliutong = _get_v(panhouliutong_raw)
+                qianzongguben = _get_v(qianzongguben_raw)
+                houzongguben = _get_v(houzongguben_raw)
+
+
+
             pos += 16
-
-            def _get_v(v):
-                if v == 0:
-                    return 0
-                else:
-                    return get_volume(v)
-
 
             row = OrderedDict(
                 [
@@ -69,12 +124,46 @@ class GetXdXrInfo(BaseParser):
                     ('month', month),
                     ('day', day),
                     ('category', category),
-                    ('panqianliutong', _get_v(panqianliutong_raw)),
-                    ('panhouliutong', _get_v(panhouliutong_raw)),
-                    ('qianzongguben', _get_v(qianzongguben_raw)),
-                    ('houzongguben', _get_v(houzongguben_raw)),
+                    ('name', self.get_category_name(category)),
+                    ('fenhong', fenhong),
+                    ('peigujia', peigujia),
+                    ('songzhuangu', songzhuangu),
+                    ('peigu', peigu),
+                    ('suogu', suogu),
+                    ('panqianliutong', panqianliutong),
+                    ('panhouliutong', panhouliutong),
+                    ('qianzongguben', qianzongguben),
+                    ('houzongguben', houzongguben),
+                    ('fenshu', fenshu),
+                    ('xingquanjia', xingquanjia)
                 ]
             )
             rows.append(row)
 
         return rows
+
+    def get_category_name(self, category_id):
+
+        if category_id in XDXR_CATEGORY_MAPPING:
+            return XDXR_CATEGORY_MAPPING[category_id]
+        else:
+            return str(category_id)
+
+
+
+
+if __name__ == '__main__':
+
+    from pytdx.util.best_ip import select_best_ip
+    from pytdx.hq import TdxHq_API
+    api = TdxHq_API()
+    with api.connect():
+        # 11 扩缩股
+        print(api.to_df(api.get_xdxr_info(1, '600381')))
+        # 12 非流通股缩股
+        #print(api.to_df(api.get_xdxr_info(1, '600339')))
+        # 13 送认购权证
+        #print(api.to_df(api.get_xdxr_info(1, '600008')))
+        # 14 送认沽权证
+        #print(api.to_df(api.get_xdxr_info(0, '000932')))
+