+++ pytdx/reader/__init__.py
@@ -1,6 +1,9 @@
 from pytdx.reader.daily_bar_reader import TdxDailyBarReader, TdxFileNotFoundException, TdxNotAssignVipdocPathException
 from pytdx.reader.min_bar_reader import TdxMinBarReader
 from pytdx.reader.lc_min_bar_reader import TdxLCMinBarReader
+from pytdx.reader.exhq_daily_bar_reader import TdxExHqDailyBarReader
+from pytdx.reader.gbbq_reader import GbbqReader
+from pytdx.reader.block_reader import BlockReader
 
 __all__ = [
     'TdxDailyBarReader',
@@ -8,4 +11,7 @@ __all__ = [
     'TdxNotAssignVipdocPathException',
     'TdxMinBarReader',
     'TdxLCMinBarReader',
+    'TdxExHqDailyBarReader',
+    'GbbqReader',
+    'BlockReader',
 ]