+++ ChangeLog.md
@@ -1,6 +1,90 @@
+1.42
+------
+* 增加了api.get_traffic_stats 获取当前连接的流量统计情况
+
+1.41
+------
+* 增加GetBlockInfoMeta， 增加GetBlockInfo， get_and_parse_block_info 等支持直接下载并解析板块文件， 感谢 @mi-fox帮助分析协议
+
+1.40
+------
+* 增加BlockReader用于读取板块列表
+
+1.39
+------
+* 修复GetSecurityQuotesCmd里面一个保留位长度计算错误导致某些行情无法获取的问题。
+
+1.38
+------
+* 修复抛出在auto_retry开启时TdxFunctionCallError异常时的一个小问题
+
+1.37
+------
+* 增加自动重连机制 auto_retry=True 时生效，并可以自定制重连策略
+* 增加连接和接口调用时的异常处理
+
+1.36
+------
+* fix bugs that break on call get_security_quotes on py2.7
+* 修复 get_security_quotes 在没有行情的时候报错 https://github.com/rainx/pytdx/issues/44
+
+1.35
+------
+* 修复 exhqapi.get_transaction_data里的增仓数据不正确, 性质数据对不上 thx @zzeric
+see https://github.com/rainx/pytdx/issues/31
+
+1.34
+------
+* 增加了 https://github.com/rainx/pytdx/issues/38 IP寻优的简单办法 
+* xdxr https://github.com/rainx/pytdx/issues/37 修改了对 11, 12, 13, 14 类别的支持， 针对13，14，增加了 fenshu xingquanjia 字段
+
+
+1.33
+------
+* 修复 Python2.x 无法安装的问题 https://github.com/rainx/pytdx/issues/36
+
+1.32
+------
+* merge pr https://github.com/rainx/pytdx/pull/34
+* 增加对 get_security_bars 的注释：如果一个股票当天停牌，那天的K线还是能取到，成交量为0
+* 上线了trade模块，但是trade模块需配合TdxTradeServer使用（一个Windows 上的 C++ 开发的Server端）
+
+1.31
+------
+* 修复除息除权信息错误，增加 fenhong  peigujia  songzhuangu  peigu suogu 字段 https://github.com/rainx/pytdx/issues/8
+
+1.30
+------
+* 修复 exhqapi.get_transaction_data里的增仓数据不正确, 性质数据对不上 https://github.com/rainx/pytdx/issues/31
+
+1.29
+------
+* 添加对通达信gbbq文件的解析类GbbqReader
+* 可以使用 hqreader -d gbbq _path_to_gbbq_file_ -o some_output_file.csv 保存股本变迁数据
+
+1.28
+------
+* 添加 get_history_instrument_bars_range 接口，可以根据一个时间范围下载历史k线信息 https://github.com/rainx/pytdx/pull/28 mifox
+
+1.27
+------
+* commit 新的get_k_data , 支持任意时间段 任意时间长度 任意股票 https://github.com/rainx/pytdx/pull/27 yutiansut
+
+1.26
+------
+* 增加了 KLINE_TYPE_EXHQ_1MIN = 7 用于获取扩展行情分钟K线 via mi-fox 
+
+1.25
+------
+* 修复了get_k_data 因为停牌导致的时间索引错位的问题 https://github.com/rainx/pytdx/pull/26 yutiansut
+
+1.24
+------
+* 增加 TdxExHqDailyBarReader ，用于读取扩展行情（如期货，现货，期权等）的盘后日线数据 https://github.com/rainx/pytdx/issues/25
+
 1.23
 ------
-增加历史分时行情，分时成交，历史分时成交 https://github.com/rainx/pytdx/pull/24 wopalm
+* 增加历史分时行情，分时成交，历史分时成交 https://github.com/rainx/pytdx/pull/24 wopalm
 
 1.22
 ------